Add post-processing to the grpc-stubs project
This project contains gRPC definitions (.proto) and generates the stubs and skeletons using the regular gRPC tooling (so not the Quarkus one). Thus, it didn't apply the post-processing steps fixing the @Generated annotation (to avoid `javax`) and removing the "final" modifiers.

This commit calls the post-processor so the generated files using the maven exec plugin.