ArC: fix dependency injection order
In case of class beans, dependency injection order is specified
by the AtInject specification as follows:

- injection points on superclasses are injected before injection points
  on subclasses;
- fields are injected before initializer methods.

Since we discover injection points bottom up, this requires reordering
after each class in the hierarchy is scanned for injection points.

The specification also prescribes that overridden initializer methods
are not injected at all. Since we discover injection points bottom up,
this requires remembering already seen methods and skipping methods
that are overridden by those that were previously seen.

This requires proper override detection. The `Methods.isOverridden()`
and `Methods.MethodKey` utils are not enough, because they actually
do _not_ detect overrides. They are used to skip processing a method
in case a method with the same signature has already been processed.
To do detect overridden methods properly, it is important to remember
that:

- private methods cannot be overridden,
- package-private methods can only be overridden in the same package.

For that reason, this commit introduces a dedicated override detection
utility.