Fix maven BuildIT failures
The dependency `stax-api` is not compatible with the dependency `quarkus-jaxb` because both add the same type `javax.xml.namespace.QName`. This was done on purpose to validate the `parent-first-artifacts` property when building projects with conflicts with Maven, so we need to skip the JaxbContext validation at build time.

This pull request should fix all the CI failures in the latest pull requests.