Switch to the new ORM 6 dialects when possible
I think we need to discuss how we will handle that as I don't think we
want a regression in features.
This PR is more for documentation than actual changes, preparing the
work to set the version in the configuration as we will probably need to
set a default minimum version for some of the dialects.