ArC: skip intercepting method calls during bean injection
Previously, method interception was skipped only during bean instance
creation. With this commit, method interception is skipped during
the entire injection process (that is, during bean constructor
invocation and initializer methods invocations).

This is the easiest way to prevent interception of initializer methods
called by the container, as demanded by the CDI specification. It also
makes sure the following provision of the Interceptors specification
is honored:

> With the exception of `@AroundConstruct` lifecycle callback interceptor
> methods, no interceptor methods are invoked until after dependency
> injection has been completed on both the interceptor instances and
> the target.