Make ORM 5 compatibility test work on machines with different timezones
Since ZoneOffsetDateTime serialization/deserialization in Hibernate ORM
5 gives different results based on the default JVM timezone, it is very
important that we use the same TZ when generating data (which happened
on my local machine) and running tests (which happens on CI).

Here I'm forcing the default TZ when generating data too, so that other
people can re-generate data even if their timezone is different.