ArC: fix inheritance of interceptor binding annotations
Class-level interceptor bindings declared on a superclass should
only be taken into account when they are `@Inherited`. ArC however
inherits all interceptor bindings from superclasses, even if
they are _not_ declared `@Inherited`. This commit fixes that in
the strict mode, to not affect Quarkus security (which assumes
that security annotations are inherited, even if they are not).

Method-level annotations are never inherited. This means that the lowest
override of a method determines whether method-level interceptor binding
annotations are present. ArC used to behave differently. If a method
on a class had no interceptor binding annotations but overrode a method
from a superclass that _did_ have interceptor binding annotations,
the method-level annotations from the superclass were inherited.
This commit fixes that.