ArC: improve equality of generated annotation literals
Apparently, a lot of people use the construct `new AnnotationLiteral<MyAnnotation>() {}`
to create an annotation literal for a memberless annotation. This is wrong, because
the result doesn't implement the annotation interface, as specified by the contract
of the `Annotation.equals()` method:

> Returns true if the specified object represents an annotation that is logically
> equivalent to this one. In other words, returns true if the specified object
> is an instance of the same annotation interface as this instance, all of whose
> members are equal to the corresponding member of this annotation, as defined below:
> [...]

Yet, this occurs both in ArC tests and in the CDI TCK. This commit therefore
improves the generated `equals()` method of ArC annotation literals for this
particular case. If the annotation type has no members, the generated `equals()`
method no longer verifies that the other object's class implements the annotation
interface. Instead, it verifies that the other object's class implements
`java.lang.annotation.Annotation` and that the `annotationType()` classes
are equal. This is also what CDI's `AnnotationLiteral` does.