Dont build if cache-hit occurs on push to fork main, the cache wont be updated.
The scheduled job means a new cache is created 10mins into the month. Due to the monthly fixed key, a cache hit will occur on every push to main in the same month after the cache is populated. The cache will not be updated if there was a hit, even if the local repo contents have changed, so running the build does nothing in terms of refreshing the cache.