Allows JSON-RPC method to return Uni<?> and add support for @Blocking and @NonBlocking
Now, JSON-PRC methods not annotated with @NonBlocking and not returning a Uni are invoked on a worker thread.
Methods annotated with @Blocking are also called on a worker thread, regardless the return type.
Methods annotated with @NonBlocking are called on the event loop even if the method does not return a Uni.