ArC: fix build compatible extensions
All the code in ArC that supports executing build compatible extensions
is present in the `processor` module, which is only available at build
time. There's however one class that needs to be present at runtime:
`ParametersImpl`, the implementation of CDI's `Parameters`.

This commit moves the class from `processor` to `runtime`.