ArC: fix spying on intercepted beans
When instantiating a `Subclass` for a bean with intercepted methods,
all `InterceptedMethodMetadata` for all intercepted methods are also
instantiated. These metadata used to hold a `Function` instance that
takes an `InvocationContext` and calls the original method after all
interceptors have executed. The `Function` instance used to remember
the `this` instance on which the method should be invoked.

This is incompatible with the way how Mockito implements spying on
objects. To create a spy, Mockito takes an object, obtains its class,
creates a subclass, and creates a copy of the original object, except
it's an instance of the subclass. When calling an intercepted method
on the spy object, the `Function` instance mentioned above diverts
the ultimate method call to the object it remembered, instead of
using the target object stored in the `InvocationContext`.

This commit changes the `InterceptedMethodMetadata` to no longer hold
a `Function`; it holds a `BiFunction` instead. The `BiFunction` is
completely stateless. Instead, the caller of the `BiFunction` must
pass not only the `InvocationContext`, but also the target object.

This way, if an intercepted object is spied upon, all intercepted
methods are eventually called correctly on the spy, instead of the
original instance.