Sanitize the behavior of provided ExceptionMapper classes in dev-mode
User provided ExceptionMapper classes now takes precedence
over the Quarkus provided NotFoundExceptionMapper.

This is done in a way that does not change anything but this
specific behavior, everything else concerning exception handling
continues to work as before.
Note also that this does not change the spec compliance in
any way

Fixes: #7883