Provide correct generic type and annotations in ParamConverterProvider
These changes will provide the correct generic type and array of annotations when the JAX-RS annotation is present on a field or a method of a Bean Param class.

The solution is similar to what was being done for the parameters of a REST Client method: it will load the metadata (generic type and annotations from fields and methods) of a BeanParam class using reflection only if there is a ParamConverterProvider present.

Fix https://github.com/quarkusio/quarkus/issues/32765

(cherry picked from commit a3d828e5a498058d6449f6f08ccd9a3fe4f1a8ed)