ArC: fix interceptor method nesting for @PostConstruct/@PreDestroy interceptors
Previously, `@PostConstruct`/`@PreDestroy` methods declared on the target
class were not invoked from within the `@PostConstruct`/`@PreDestroy`
methods declared on an interceptor class. Instead, they were invoked
_after_ the interceptor chain finished, effectively creating two chains
of lifecycle callback interceptors.

This didn't allow an "outer" lifecycle callback method, when declared
on an interceptor class, to catch and handle an exception thrown by an
"inner" lifecycle callback method, when declared on the target class.

This commit fixes the implementation of `@PostConstruct`/`@PreDestroy`
interceptors by reifying the "inner" chain of lifecycle callback methods,
declared on the target class, into a `Runnable` and passing that to the
"outer" chain of lifecycle callback methods, declared on the interceptor
class, to be invoked at the very end of the chain. This effectively
merges the two chains into one, as prescribed by the specification.

Note that the "inner" chain is reified into a `Runnable` only when
an "outer" chain actually exists. It often doesn't, in which case,
allocating a `Runnable` only to call it is a waste of resources.