ArC: fix bean creation in presence of @AroundConstruct interceptors
`@AroundConstruct` interceptors may modify the constructor arguments
using `InvocationContext.setParameters()`. ArC used to ignore this
modification and call the bean constructor using the original
arguments, obtained before calling the interceptors. This commit
fixes that and uses the arguments array from `InvocationContext`
to call the bean constructor.

To do that, the forwarding function passed along the interceptor
chain is no longer a `Supplier<Object>` (which used to call the bean
constructor with original arguments), but a `Function<Object[], Object>`
(which is passed the arguments stored in the `InvocationContext` and
uses them to call the bean constructor).