ArC: fix generated Bean.destroy() in case a user directly passes in a client proxy
ArC callers of `Bean.destroy()` never pass in a client proxy, but
there may be direct callers too, even though this API is low-level.
This commit unwraps the client proxy if it was passed to `destroy()`
before proceeding, which avoids a `ClassCastException` in case
the bean has a `@PreDestroy` interceptor (and hence the contextual
instance is a generated subclass).