Fail on attempt to set custom hostname verifier
Both custom HostnameVerifier and SSLContext are currently not
supported by resteasy-reactive-client. This is documented as
known limitations.

This change aligns the implementation to the already present state
for SSLContext. It makes the builder fail when HostnameVerifier is
tried to be configured. The benefit is that it fails as early as
currently possible and not only when the custom hostname verification
would be needed later on.

Switch both to UnsupportedOperationException.

On the rest-client-reactive settings a custom hostname
verifier is currently not supported. Adapted the tests
accordingly as the builder fails early now throwing
UnsupportedOperationException.