ArC: treat additional scope annotations correctly, even if there's no corresponding context
It makes no sense to register an additional scope annotation without
registering a corresponding context, but the CDI TCK uses that and
it isn't terribly hard to support.

With this commit, ArC treats additional bean defining annotations
that are meta-annotated `@Scope` or `@NormalScope` as scope annotations,
even if there is no corresponding context.

This commit also fixes the type discovery implementations to also discover
scope annotations and register them as additional bean defining annotations.