Fix the Micrometer MP Metrics annotation transformation
The annotation transformation replaces an existing annotation with
a new one without paying attention to the annotation's target. That's
usually OK, because the `TransformationContext` typically contains
the correct target, but that isn't true for method parameter annotations.
For legacy reasons, the annotation transformation API treats them as
annotations on methods, so we need to pay extra attention to preserve
the annotation target. This commit does that.