ArC: fix bean archive discovery in strict mode
The Quarkus ArC extension assumes that the root application archive
is always a bean archive. That doesn't necessarily have to be true
per the CDI specification, but is useful in practice. Therefore, it
needs to be fixed only in strict mode, which is what this commit does.

By default, the root application archive is always a bean archive.
In strict mode, the root application archive goes through the same
scanning process as other application archives.