Arquillian: add workaround for memory leak
Arquillian holds onto deployment-scoped (and container-scoped) objects
until its end of life. That allows undeploying and redeploying without
losing contextual data, but is also technically a memory leak. In case
the deployment-scoped objects are large (which is the case in Quarkus,
see `RunningQuarkusApplication`), this memory leak becomes very visible
very soon.

The Quarkus Arquillian adapter is only used to run TCKs for Jakarta
and MicroProfile specifications, which don't need this capability. Also,
the `RunningQuarkusApplication` is unusable after it is closed, so
it makes no sense to keep it around. Hence, `QuarkusDeployableContainer`
no longer stores the `RunningQuarkusApplication` into the deployment
context directly; instead, it stores a holder object (`QuarkusDeployment`)
that can safely be cleaned up after an Arquillian deployment is undeployed.
That makes the Quarkus application unreachable and GC-able. Arquillian
will keep piling up empty objects, which is still a memory leak, but
one that doesn't hurt as much.

This commit also fixes an ordering issue. Previously, the Arquillian
adapter used to delete the Quarkus application directory _before_ shutting
down the application, which could lead to errors. This commit switches
the order: the app is shut down first and only then is it deleted.