Testing: improve instantiating test classes through CDI
This commit moves the code that looks up a test class from CDI into
a dedicated delegate class. This class is loaded from the Quarkus CL
and so the code can be written directly instead of reflectively.

Additionally, the CDI lookup can now correctly instantiate a generic
test class. Such bean cannot be looked up from CDI directly, because
its set of bean types doesn't contain a `Class` object (it contains
a `ParameterizedType` instead). This commit uses a naive approach
for this (iterate through all beans), which is fine, because it's
very uncommon.