Allows receiving the channel name when using Redis Pub/Sub
When using Redis Pub/Sub, receiving the message channel was impossible.
This was a problem when subscribing to multiple channels or using regexp.

This commit adds methods for receiving a bi-consumer to get the channel name and payload.
When not possible (for method returning a Multi), a dedicated structure is introduced (RedisPubSubMessage).