ArC: speed up bean resolution at runtime
Specifically, the invocation of BeanManaget#getBeans() and the first hit
for any programmatic lookup.

We use a precomputed map of a raw type to the set of matching beans
which reduces the number of beans we need to iterate over when we're
looking for matching beans.

Note that BeanManaget#getBeans() does not use the computing cache of
results so this optimizaton results in a significant speed-up. On the
other hand, the precomputed map could be memory demanding for large
applications.