Remove unnecessary --pull parameter when running builder images
Follows up on #33749.

I forgot to remove this code that was just a remainder of an earlier
approach which didn't go anywhere
(https://github.com/quarkusio/quarkus/issues/33691#issuecomment-1570453108).

Keeping this code could in theory lead to extra pulls when using
quarkus.native.build-image.pull='always' (the default) and the image
gets updated right while we're building, but more importantly it could
lead to extra, unecessary queries to quay.io, so let's avoid that.