ArC: prevent removing the final flag from records
Record classes must always be `final`, even the `Class.isRecord()`
method checks that. Therefore, the bytecode transformation that
removes the `final` flag from a class is invalid for records.

With this commit, ArC will no longer blindly remove the `final`
flag from records; a deployment problem occurs instead. This means
that records cannot be used to define normal scoped beans and
cannot be intercepted or decorated.