Fix parameter validation issue in classic rest-client
URL parsing inside of `QuarkusRestClientBuilder.verifyInterface` used
`ResteasyUriBuilderImpl.uri`, which works for most cases, but parsed in
subtly wrong ways when the constructed path started with `//`. For
example, a path of `//{id}` parses `{id}` as the hostname. Such a path
was possible to trigger by a couple of scenarios:

1. Interface has `@Path("")`, and a method-level `@Path` value starting
   with `/`.
2. Interface has `@Path("/")`, and a method-level `@Path` value not
   starting with `/`.

Furthermore, if the first path component contained a path param, then it
would not be counted, leading to incorrect validation results.

Added an empty path test to rest-client-reactive.