ArC: add support for custom AlterableContext implementations
Custom contexts in ArC must implement the `InjectableContext`
interface, which is ArC-specific. To implement CDI Lite properly,
ArC must also support custom implementations of `AlterableContext`.

Fortunately, the `InjectableContext` interface adds just a few
methods on top of `AlterableContext`, and none of them are critical
for custom contexts to function. Therefore, with this commit, we
simply take the user-supplied implementation of `AlterableContext`
and generate a subclass that implements `InjectableContext`, where
all the additional methods throw `UnsupportedOperationException`.
This subclass is then registered as the custom context, instead
of the original user-supplied class.