Fix span export issue when user provided SpanProcessor beans exist
SpanProcessor implementations are hooked into OTel in Quarkus
via the AutoConfiguredOpenTelemetrySdkBuilder#addTracerProviderCustomizer
mechanism, which in turn obtains those processor from CDI.
However, because LateBoundBatchSpanProcessor (which is the one
that ends up being responsible for the span export
via its BatchSpanProcessor delegate) was marked as a
@DefaultBean, when a user provided SpanProcessor became
part of the application, the LateBoundBatchSpanProcessor
was completely ignored.
The fix is simply to remove @DefaultBean and always have
LateBoundBatchSpanProcessor be considered as a SpanProcessor.

Fixes: #33407
(cherry picked from commit bfecf760ca4309444213748a3b9c62f9a5badcb0)