Fail when GraalVM / Mandrel version detection fails
It's easy to miss the error message and Quarkus will assume the version
is a snapshot of GraalVM making it harder to debug.

The benefits of supporting arbitrary versions from dev/snapshot builds
doesn't seem to justify the above so it's better to just fail and
require a proper version in dev/snapshot builds.