Implement support for @RunOnVirtualThread for gRPC services.
This commit allows gRPC methods to be invoked on virtual threads (while keeping the duplicated context and request scope). A new virtual thread is created for each invocation.

Note that methods ingesting streams are not supported for now.

This commit also sets up the structure to run integration tests verifying the virtual thread support.