Add `maxLifetime` to reactive datasource configurations
This enables configuration of the new`maxLifetime` value provided by Vert.x 4.3 that ensures connections are recycled after a maximum duration.

Because `maxLifetime` could conceivable be set beyond the maximum value that milliseconds can store in an `int` (24.8 days) I created a `UnitisedTime` class. The `UnitisedTime` class has a `unitised` method that converts a `Duration` into the smallest possible `TimeUnit` that can be stored in an int, starting with milliseconds.

For consistency, I updated the `idleTimeout` to use the same `UnitisedTime.unitised` utitlity for conversion.