kafka-streams: restore the feature name at Quarkus startup
Fixes #34171

The FeatureBuildItem was returned in a BuildStep now annotated to only run when native compilation is used.
The consequence is that in JVM mode, the FeatureBuildItem is not created, and the feature name not appearing in the start logs of Quarkus.

Solution: extract the FeatureBuildItem creation in a dedicated BuildStep.
(cherry picked from commit b884a10acdd51740ab2e1aed015e3b9e3d3d9ef5)