Avoid keeping references to BytecodeRecorderImpl
BytecodeRecorderImpl references are kept around for the whole life of
the class loader. And these classes can be huge.

Moving the static map to another class so that is it properly isolated
and made sure we used a proper static access in the close task.

(cherry picked from commit cb33a01a2e8b64076dd1cb31a07d6024885b2684)