Rework Flyway CDI initialization
In passing, made the Flyway beans singleton as we access them a lot more
via CDI and we would have created a lot more beans.
Given the comment I removed where the Dependent scope was added, I think
we can consider it is a good move.
I'm aware this means, we will keep Flyway beans around but that's the
price we have to pay to not slow down the initialization too much.