refactor: OpenShiftClientProducer reuses the KubernetesClient default bean
Instead of creating a new bean (which might lead to an application having two
separate KubernetesClient instances) the OpenShiftClientProducer now
relies on the default KubernetesClientProducer default bean and adapts
it.

To avoid circular dependency and multiple bean candidate issue I've used
a `@Named` annotation to qualify the default KubernetesClient bean.

However, this might be problematic in the scenario where a provides
a different base KubernetesClient bean. In this case, unless that bean
is qualified too, it won't be reused. This is the exact same situation
we have now, so I'm assuming it won't be a problem. Anyway, if there's
a better solution that covers all cases, then we might want to implement it
here (please let me know).

Signed-off-by: Marc Nuri <marc@marcnuri.com>