Significantly improve the output of the update command
When having an error, it was extremely confusing, you had confusing
error messages, a lot of nested exceptions...

This will improve the output, make sure we have proper colors for the
log levels and avoid having two levels logged.
Also avoid a couple of nested exceptions and make the error messages
more actionable.