Separate Hibernate Reactive tests against MariaDB / MySQL
This creates a module integration-tests/hibernate-reactive-mariadb next
to integration-tests/hibernate-reactive-mysql,
and updates integration-tests/hibernate-reactive-mysql* to actually use
a MySQL instance (and not MariaDB like it used to).

This is important because Hibernate Reactive has a different dialect for
MariaDB and MySQL, so tests passing with one may not pass with the
other.