Move the loom-unit extension as an independent project
This commit brings the functionality from loom-unit (partially, but the important ones are there) into Quarkus.
Thus, @ShouldPin and @ShouldNotPin can be used on a @QuarkusTest directly as soon as the junit5-virtual-threads dependency is in the project and the extension declared on the test case with the `io.quarkus.test.junit5.virtual.VirtualThreadUnit` annotation.

loom-unit original repo: https://github.com/cescoffier/loom-unit