Check for code usage fluctuations in native images
This aims to aid in detecting code usage fluctuations as early as
possible. I set the threshold to 3% which might be a bit tight.

Ideally most (if not all) ITs being tested in native mode should include
such checks. This way when a dependency gets updated and brings in more
bloat, or a GraalVM change results in more code becoming reachable we
will be able to notice.

A potential drawback is that for these tests to work they require
`quarkus.native.enable-reports` to be set, which might not be desirable.